fix：env改进

+++ NChat-React/n-chat/src/components/Logout.jsx
@@ -3,12 +3,21 @@ import { useNavigate } from "react-router-dom"
 import { Button } from "antd"
 import { PoweroffOutlined } from "@ant-design/icons"
 import styled from "styled-components"
+import axios from "axios";
+import { logoutRoute } from "../utils/APIRoutes";
+
 
 export default function Logout() {
 	const navigate = useNavigate()
 	const handleClick = async () => {
-		localStorage.clear()
-		navigate("/login")
+		const id = await JSON.parse(
+      localStorage.getItem(process.env.REACT_APP_LOCALHOST_KEY)
+    )._id;
+		const data = await axios.get(`${logoutRoute}/${id}`);
+    if (data.status === 200) {
+      localStorage.clear();
+      navigate("/login");
+    }
 	}
 	return (
 		<ButtonContainer>
